Move JWT creation code to models
